backend/requirements.txt
# Core Backend Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Data Processing
pandas>=2.1.0
numpy>=1.26.0

# Machine Learning & Forecasting
prophet>=1.1.5
statsmodels>=0.14.0
scikit-learn>=1.3.0
xgboost>=2.0.0

# HTTP & API
requests>=2.31.0
httpx>=0.25.0

# Configuration & Validation
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
supabase>=2.3.0

# PDF Reports & Parsing
reportlab>=4.0.0
pdfplumber>=0.11.0

# Testing
pytest>=7.4.0

# SHAP for explainability (optional - may not work on Render free tier)
# shap>=0.44.0

# Local AI (TinyLlama)
torch>=2.0.0
transformers>=4.36.0
accelerate>=0.25.0
# bitsandbytes>=0.41.0  # Optional: 4-bit quantization (GPU only)

# Research Agent
tavily-python>=0.3.0
feedparser>=6.0.0

# Background Tasks
celery>=5.3.0
# redis>=5.0.0  # Optional: use Redis broker instead of SQLite

# Database (Newsletter)
sqlalchemy>=2.0.0
alembic>=1.13.0

# Email Templates
jinja2>=3.1.0
